# GymApi.Api/openapi/session-tracking.yaml
openapi: 3.0.3
info:
  title: GymApi – Current Session Tracking
  description: Core subdomain for recording an active training session in real time.
  version: "1.0.0"

servers:
  - url: https://api.gymapp.example.com
    description: Production (Azure Container Apps)
  - url: http://localhost:5000
    description: Local development

tags:
  - name: Sessions
    description: Training session lifecycle

paths:

  /api/sessions:
    post:
      tags: [ Sessions ]
      operationId: createSession
      summary: Create a new training session
      description: |
        Optionally inherits exercises from a previous session as **pending templates**
        (label, photo, properties are copied — times are not).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              userId: "a1b2c3d4-0000-0000-0000-000000000001"
              inheritFromSessionId: "a1b2c3d4-0000-0000-0000-000000000000"
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/sessions/{sessionId}:
    get:
      tags: [ Sessions ]
      operationId: getSession
      summary: Get a session by ID
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/sessions/{sessionId}/exercises:
    post:
      tags: [ Sessions ]
      operationId: addExercise
      summary: Add and immediately start a new exercise
      description: |
        Creates a new exercise entry and starts it (`status: Running`).
        If another exercise is currently running it is **automatically finished** first.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddExerciseRequest'
            example:
              autoLabel: "Bench Press"
              photoUrl: "https://cdn.example.com/exercises/bench-press.jpg"
              maxEndAt: "2024-06-01T10:15:00Z"
              properties:
                - name: "Weight"
                  value: "80 kg"
                - name: "Reps"
                  value: "8"
      responses:
        '201':
          description: Exercise added and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/sessions/{sessionId}/exercises/{exerciseId}/start:
    post:
      tags: [ Sessions ]
      operationId: startExercise
      summary: Start a pending (inherited) exercise
      description: |
        Starts an exercise that was cloned from a previous session as a pending
        template (`status: Pending → Running`).
        Any currently running exercise is **automatically finished** first.
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/ExerciseId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExerciseRequest'
      responses:
        '200':
          description: Exercise started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExerciseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/sessions/{sessionId}/exercises/{exerciseId}:
    delete:
      tags: [ Sessions ]
      operationId: removeExercise
      summary: Remove an exercise from an active session
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/ExerciseId'
      responses:
        '204':
          description: Exercise removed
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/sessions/{sessionId}/finish:
    post:
      tags: [ Sessions ]
      operationId: finishSession
      summary: Finish the training session
      description: |
        Transitions the session to `Finished`.
        Any running exercise is **automatically finished** first.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

components:

  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExerciseId:
      name: exerciseId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Operation invalid for the current aggregate state (e.g. session already finished)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    # ── Requests ──────────────────────────────────────────────────────────────

    CreateSessionRequest:
      type: object
      required: [ userId ]
      properties:
        userId:
          type: string
          format: uuid
          description: Authenticated user creating the session.
        inheritFromSessionId:
          type: string
          format: uuid
          nullable: true
          description: |
            When provided, exercises from that session are cloned as pending
            templates into the new session (label, photo, properties — no times).

    AddExerciseRequest:
      type: object
      required: [ autoLabel ]
      properties:
        autoLabel:
          type: string
          minLength: 1
          description: System-generated or user-confirmed exercise name.
          example: "Bench Press"
        photoUrl:
          type: string
          format: uri
          nullable: true
          example: "https://cdn.example.com/exercises/bench-press.jpg"
        maxEndAt:
          type: string
          format: date-time
          nullable: true
          description: Optional hard deadline used as an in-app timer / alarm.
        properties:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ExercisePropertyDto'

    StartExerciseRequest:
      type: object
      properties:
        maxEndAt:
          type: string
          format: date-time
          nullable: true
          description: Optional hard deadline for this exercise run.

    ExercisePropertyDto:
      type: object
      required: [ name, value ]
      properties:
        name:
          type: string
          example: "Weight"
        value:
          type: string
          example: "80 kg"

    # ── Responses ─────────────────────────────────────────────────────────────

    SessionResponse:
      type: object
      required: [ id, userId, createdAt, status, exercises ]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ Active, Finished ]
        inheritedFromSessionId:
          type: string
          format: uuid
          nullable: true
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseResponse'

    ExerciseResponse:
      type: object
      required: [ id, autoLabel, status, properties ]
      properties:
        id:
          type: string
          format: uuid
        autoLabel:
          type: string
        photoUrl:
          type: string
          format: uri
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
          description: Null when status is Pending.
        maxEndAt:
          type: string
          format: date-time
          nullable: true
        realEndAt:
          type: string
          format: date-time
          nullable: true
          description: Null until the exercise is finished.
        status:
          type: string
          enum: [ Pending, Running, Finished ]
          description: |
            - **Pending** – inherited template, not yet started
            - **Running** – started, no end time yet
            - **Finished** – realEndAt is set
        properties:
          type: array
          items:
            $ref: '#/components/schemas/ExercisePropertyDto'

    ErrorResponse:
      type: object
      required: [ error ]
      properties:
        error:
          type: string